Navbar logic fixed
