+++ src/components/ErrorPages/404.js
@@ -1,12 +1,7 @@
 import React from "react";
-import MiniNavbar from "../MiniNavbar";
 
 const NotFound = () => {
-  return (
-    <div>
-      <MiniNavbar /> Whoopsie Daisy!!! Page not found
-    </div>
-  );
+  return <div>Whoopsie Daisy!!! Page not found</div>;
 };
 
 export default NotFound;
