+++ src/components/NavBar/navbarPaths.js
@@ -1,2 +1,2 @@
 // These are the routes that the main navbar should be shown
-module.exports = ["/dashboard"];
+module.exports = ["/login", "/signup"];
